third_party/blink/web_tests/fast/dom/CSSStyleDeclaration/css-style-declaration-named-setter-expected.txt

PASS document.body.style.margin is "0px"
PASS document.body.style.margin = 1 is 1
PASS document.body.style.margin is "1px"
PASS document.body.style.margin = badString threw exception Exception in toString().
PASS document.body.style.margin is "1px"
PASS document.body.style.__proto__.margin = 2 is 2
PASS document.body.style.__proto__.margin is 2
PASS document.body.style.margin is 2
PASS document.body.style.margin = 3 is 3
PASS document.body.style.margin is 3
PASS successfullyParsed is true

TEST COMPLETE
